requirements.txt
Django==4.2.7
django_ckeditor==6.7.0
python-decouple==3.8
